pyproject.toml
[project]
name = "medical-ai-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.120.4",
    "langchain>=1.0.3",
    "langchain-google-genai>=3.0.1",
    "pandas>=2.3.3",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.3",
    "python-socketio[asyncio]>=5.14.3",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-to-pydantic>=0.0.8",
    "tabulate>=0.9.0",
    "uvicorn[standard]>=0.38.0",
]
